refactor(cli): extract runtime repl
