+++ packages/agent-cli/agent-cli.cabal
@@ -56,6 +56,7 @@ library
                     , Agent.CLI.Runtime.Internal
                     , Agent.CLI.Runtime.Persistence
                     , Agent.CLI.Runtime.Recap
+                    , Agent.CLI.Runtime.Repl
                     , Agent.CLI.Runtime.Types
                     , Agent.CLI.Session.Attachments
                     , Agent.CLI.Session.Choices
